portafolio_app/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-jazzmin==3.0.0
idna==3.7
mysqlclient==2.2.4
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
djangorestframework==3.15.2
